Move scanner struct to header
