+++ boreas/alivedetection.c
@@ -75,32 +75,6 @@ get_alive_test_methods (alive_test_t *alive_test);
 static boreas_error_t
 set_socket (socket_type_t socket_type, int *scanner_socket);
 
-/**
- * @brief The scanner struct holds data which is used frequently by the alive
- * detection thread.
- */
-struct scanner
-{
-  /* sockets */
-  int tcpv4soc;
-  int tcpv6soc;
-  int icmpv4soc;
-  int icmpv6soc;
-  int arpv4soc;
-  int arpv6soc;
-  /* UDP socket needed for getting the source IP for the TCP header. */
-  int udpv4soc;
-  int udpv6soc;
-  /* TH_SYN or TH_ACK */
-  uint8_t tcp_flag;
-  /* ports used for TCP ACK/SYN */
-  GArray *ports;
-  /* redis connection */
-  kb_t main_kb;
-  /* pcap handle */
-  pcap_t *pcap_handle;
-};
-
 /* Max_scan_hosts and max_alive_hosts related struct. */
 struct scan_restrictions
 {
